+++ core/prelude-custom.el
@@ -36,6 +36,13 @@
   :prefix "prelude-"
   :group 'convenience)
 
+(defcustom prelude-minimalistic-ui nil
+  "Controls whether to display the menu-bar and line numbers.
+Note that the toolbar is always hidden regardless of this setting."
+  :type 'boolean
+  :group 'prelude
+  :package-version '(prelude . "1.1"))
+
 (defcustom prelude-auto-save t
   "Non-nil values enable Prelude's auto save."
   :type 'boolean