+++ core/prelude-ui.el
@@ -37,7 +37,8 @@
 (when (fboundp 'tool-bar-mode)
   (tool-bar-mode -1))
 
-(menu-bar-mode -1)
+(when prelude-minimalistic-ui
+  (menu-bar-mode -1))
 
 ;; the blinking cursor is nothing, but an annoyance
 (blink-cursor-mode -1)