+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changes
 
+* [#1278](https://github.com/bbatsov/prelude/issues/1278): Don't disable `menu-bar-mode` unless `prelude-minimalistic-ui` is enabled.
+
 ### Bugs fixed
 
 ## 1.0.0 (2020-09-15)